wikipedia_etl/requirements.txt
apache-airflow
beautifulsoup4
mysql-connector-python
hdfs
pyyaml
pytest
requests
